Add redirect url to post checkout
